Fix: 엔티티 매핑오류

See merge request s11-webmobile1-sub2/S11P12D107!101

+++ backend/src/main/java/com/chestnut/backend/qna/QnACategory.java
@@ -13,7 +13,7 @@ public class QnACategory {
 
     @Id
     @Column(columnDefinition = "tinyint", nullable = false)
-    private Integer qnaCategoryId;
+    private Byte qnaCategoryId;
 
     @Column(columnDefinition = "varchar(5)", nullable = false)
     private String qnaCategoryName;
