+++ backend/src/main/java/com/chestnut/backend/study/dto/PhonologyStudyInfo.java
@@ -11,7 +11,7 @@ import lombok.ToString;
 @AllArgsConstructor
 public class PhonologyStudyInfo {
     private String categoryContent;
-    private Long studyCategoryId;
+    private Byte studyCategoryId;
     private Long studyId;
     private String phonologyRule;
     private String phonologyExplanation;
